requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.2  # or faiss-gpu if CUDA is available
psutil>=5.8.0
pydantic>=1.8.0
python-dotenv>=0.19.0
numpy>=1.21.0  # Required by FAISS and sentence-transformers